Initial Check in of Word_Form_Data_Extract.vba
When running the Macro in Office 2010 it looks at current documents and pulle the tags of the content controls as the first row in a tab delimited txt file.  The data excluding any fields not filled in by user is added to the next line of the tab delimited file. if no data in a field a tab is added to retain all data being in correct tag name column. 

Future changes to make:
The code as written is to be added to the code of a 2010 word macro enabled document and used to scan that specific documents content controls for header and record information. 
The macro will be change d so there is the form file for the users and there is a word macro file for the administrator.
There will be a printable form file that doesn't print the field prompt txt
There will be a fillable form used as softcopy to fill and return. 
The macro will open a select folder where all the filled out and returned forms are saved. 
prompt for where process files to be placed. 
files are moved in processed folder with the following folders. RAW, Processed, Invalid
The ones in the processed file are a copy of the RAW ones with a date/time stamp on them. as each file is processed, read into memory, the file is moved to RAW and processed and renamed. 
This master macro word.docm file should be able to read any form file that follows these restraints. 
a. All content controls have a tag
b. All tags on content controls are unique

The Code as is, needs optimization.
Optimization planned for completion when updating to a master macro file that prompts for file locations