Delimiter selection implemented
dropdown selection was already on form. This added a sub to call at time of clicking process. it reads the text of the selection and then sets a global variable to vbTab or Chr(44) to be used when iterating through arrays to create output strings and to be used for setting the file extension on mydataresults file to either .txt or .csv.